python/requirements.txt
opencv-python==4.9.0.80
mediapipe==0.10.33
numpy==1.26.4
pyserial==3.5
matplotlib==3.8.2